Consolidate LLM utilities into unified handler
