+++ real-treasury-business-case-builder.php
@@ -279,11 +279,10 @@ return true;
 				   'inc/class-rtbcb-validator.php'		 => 'RTBCB_Validator',
 				   'inc/class-rtbcb-maturity-model.php'	 => 'RTBCB_Maturity_Model',
 				   'inc/class-rtbcb-leads.php'		 => 'RTBCB_Leads',
-				   'inc/class-rtbcb-rag.php'             => 'RTBCB_RAG',
+                                   'inc/class-rtbcb-rag.php'             => 'RTBCB_RAG',
                                    'inc/class-rtbcb-response-parser.php' => 'RTBCB_Response_Parser',
-                                   'inc/class-rtbcb-llm.php'             => 'RTBCB_LLM',
-                                   'inc/class-rtbcb-llm-optimized.php'   => 'RTBCB_LLM_Optimized',
-				   'inc/class-rtbcb-intelligent-recommender.php' => 'RTBCB_Intelligent_Recommender',
+                                   'inc/class-rtbcb-llm-unified.php'     => 'RTBCB_LLM_Unified',
+                                   'inc/class-rtbcb-intelligent-recommender.php' => 'RTBCB_Intelligent_Recommender',
                                    'inc/class-rtbcb-json-error.php'     => 'RTBCB_JSON_Error',
                                    'inc/class-rtbcb-response-integrity.php' => 'RTBCB_Response_Integrity',
                                    'inc/class-rtbcb-router.php'		 => 'RTBCB_Router',
