+++ tests/RTBCB_StructureReportDataTest.php
@@ -8,7 +8,7 @@ use PHPUnit\Framework\TestCase;
 
 require_once __DIR__ . '/wp-stubs.php';
 require_once __DIR__ . '/../inc/class-rtbcb-ajax.php';
-require_once __DIR__ . '/../inc/class-rtbcb-llm.php';
+require_once __DIR__ . '/../inc/class-rtbcb-llm-unified.php';
 
 if ( ! function_exists( '__' ) ) {
 function __( $text, $domain = null ) {
